+++ example/example.dart
@@ -1,5 +1,10 @@
-import 'package:flutter/cupertino.dart' show CupertinoIcons, CupertinoThemeData;
-import 'package:flutter/material.dart' show ThemeData, Colors, Icons;
+import 'package:flutter/cupertino.dart'
+    show
+        CupertinoActionSheet,
+        CupertinoActionSheetAction,
+        CupertinoIcons,
+        CupertinoThemeData;
+import 'package:flutter/material.dart' show Colors, Icons, ThemeData;
 import 'package:flutter/widgets.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
@@ -33,11 +38,11 @@ class _AppState extends State<App> {
 
     return PlatformProvider(
       builder: (BuildContext context) => PlatformApp(
-            title: 'Flutter Platform Widgets',
-            android: (_) => new MaterialAppData(theme: themeData),
-            ios: (_) => new CupertinoAppData(theme: cupertinoTheme),
-            home: LandingPage(),
-          ),
+        title: 'Flutter Platform Widgets',
+        android: (_) => new MaterialAppData(theme: themeData),
+        ios: (_) => new CupertinoAppData(theme: cupertinoTheme),
+        home: LandingPage(),
+      ),
     );
   }
 }
@@ -157,13 +162,19 @@ class LandingPageState extends State<LandingPage> {
               onPressed: () => _showExampleDialog(),
             ),
             Divider(),
-            SectionHeader(title: '4. Navigation'),
+            SectionHeader(title: '4. Popup/Sheet'),
+            PlatformButton(
+              child: PlatformText('Show Popup/Sheet'),
+              onPressed: () => _showPopupSheet(),
+            ),
+            Divider(),
+            SectionHeader(title: '5. Navigation'),
             PlatformButton(
               child: PlatformText('Open Tabbed Page'),
               onPressed: () => _openPage((_) => new TabbedPage()),
             ),
             Divider(),
-            SectionHeader(title: '5. Advanced'),
+            SectionHeader(title: '6. Advanced'),
             PlatformButton(
               child: PlatformText('Page with ListView'),
               onPressed: () => _openPage((_) => new ListViewPage()),
@@ -171,9 +182,9 @@ class LandingPageState extends State<LandingPage> {
             PlatformWidget(
               android: (_) => Container(), //this is for iOS only
               ios: (_) => PlatformButton(
-                    child: PlatformText('iOS Page with Colored Header'),
-                    onPressed: () => _openPage((_) => new ListViewHeaderPage()),
-                  ),
+                child: PlatformText('iOS Page with Colored Header'),
+                onPressed: () => _openPage((_) => new ListViewHeaderPage()),
+              ),
             ),
           ],
         ),
@@ -190,25 +201,94 @@ class LandingPageState extends State<LandingPage> {
     );
   }
 
+  _showPopupSheet() {
+    showPlatformModalSheet(
+      context: context,
+      builder: (_) => PlatformWidget(
+        android: (_) => _androidPopupContent(),
+        ios: (_) => _cupertinoSheetContent(),
+      ),
+    );
+  }
+
+  Widget _androidPopupContent() {
+    return Container(
+      padding: const EdgeInsets.all(16),
+      child: Column(
+        children: <Widget>[
+          GestureDetector(
+            child: Container(
+              padding: const EdgeInsets.all(8),
+              child: PlatformText('Option 1'),
+            ),
+            onTap: () => Navigator.pop(context),
+          ),
+          GestureDetector(
+            child: Container(
+              padding: const EdgeInsets.all(8),
+              child: PlatformText('Option 2'),
+            ),
+            onTap: () => Navigator.pop(context),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _cupertinoSheetContent() {
+    return CupertinoActionSheet(
+      title: const Text('Favorite Dessert'),
+      message:
+          const Text('Please select the best dessert from the options below.'),
+      actions: <Widget>[
+        CupertinoActionSheetAction(
+          child: const Text('Profiteroles'),
+          onPressed: () {
+            Navigator.pop(context, 'Profiteroles');
+          },
+        ),
+        CupertinoActionSheetAction(
+          child: const Text('Cannolis'),
+          onPressed: () {
+            Navigator.pop(context, 'Cannolis');
+          },
+        ),
+        CupertinoActionSheetAction(
+          child: const Text('Trifle'),
+          onPressed: () {
+            Navigator.pop(context, 'Trifle');
+          },
+        ),
+      ],
+      cancelButton: CupertinoActionSheetAction(
+        child: const Text('Cancel'),
+        isDefaultAction: true,
+        onPressed: () {
+          Navigator.pop(context, 'Cancel');
+        },
+      ),
+    );
+  }
+
   _showExampleDialog() {
     showPlatformDialog(
       context: context,
       builder: (_) => PlatformAlertDialog(
-            title: Text('Alert'),
-            content: Text('Some content'),
-            actions: <Widget>[
-              PlatformDialogAction(
-                android: (_) => MaterialDialogActionData(),
-                ios: (_) => CupertinoDialogActionData(),
-                child: PlatformText('Cancel'),
-                onPressed: () => Navigator.pop(context),
-              ),
-              PlatformDialogAction(
-                child: PlatformText('OK'),
-                onPressed: () => Navigator.pop(context),
-              ),
-            ],
+        title: Text('Alert'),
+        content: Text('Some content'),
+        actions: <Widget>[
+          PlatformDialogAction(
+            android: (_) => MaterialDialogActionData(),
+            ios: (_) => CupertinoDialogActionData(),
+            child: PlatformText('Cancel'),
+            onPressed: () => Navigator.pop(context),
           ),
+          PlatformDialogAction(
+            child: PlatformText('OK'),
+            onPressed: () => Navigator.pop(context),
+          ),
+        ],
+      ),
     );
   }
 }