+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.13.0] - September 13, 2019
+
+#### (Flutter version support: v1.9.1+hotfix.2)
+
+- Updated all widgets compatible with Flutter v 1.9.1
+- Added showPlatformModalSheet
+
 ## [0.12.0] - June 24, 2019
 
 #### (Flutter version support: v1.5.4-hotfix.2)